+++ e2e/packages/test-data/package.json
@@ -5,9 +5,9 @@
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "generate-test-data": "tsx generate-test-data.ts",
-    "generate-test-data-query": "tsx generate-test-data-query.ts",
-    "generate-test-data-records": "tsx generate-test-data-records.ts"
+    "generate-bench-data-bulk": "tsx generate-bench-data-bulk.ts",
+    "generate-bench-data-query": "tsx generate-bench-data-query.ts",
+    "generate-test-data": "tsx generate-test-data.ts"
   },
   "devDependencies": {
     "@latticexyz/block-logs-stream": "link:../../../packages/block-logs-stream",
