+++ packages/store-sync/benchmarks/getRecord.bench.ts
@@ -5,9 +5,9 @@ import { encodeEntity } from "../src/recs";
 import { logsToBlocks } from "../test/logsToBlocks";
 import { buildTable, getTables } from "../src/sqlite";
 import { eq } from "drizzle-orm";
-import worldRpcLogs10 from "../../../test-data/world-logs-10.json";
-import worldRpcLogs100 from "../../../test-data/world-logs-100.json";
-import worldRpcLogs1000 from "../../../test-data/world-logs-1000.json";
+import worldRpcLogs10 from "../../../test-data/world-logs-bulk-10.json";
+import worldRpcLogs100 from "../../../test-data/world-logs-bulk-100.json";
+import worldRpcLogs1000 from "../../../test-data/world-logs-bulk-1000.json";
 
 describe.each([
   { numRecords: 10, logs: worldRpcLogs10 },
