+++ packages/store-sync/benchmarks/storageAdapter.bench.ts
@@ -1,9 +1,9 @@
 import { bench, describe } from "vitest";
 import { createRecsStorage, createSqliteStorage, createZustandStorage } from "../test/utils";
 import { logsToBlocks } from "../test/logsToBlocks";
-import worldRpcLogs10 from "../../../test-data/world-logs-10.json";
-import worldRpcLogs100 from "../../../test-data/world-logs-100.json";
-import worldRpcLogs1000 from "../../../test-data/world-logs-1000.json";
+import worldRpcLogs10 from "../../../test-data/world-logs-bulk-10.json";
+import worldRpcLogs100 from "../../../test-data/world-logs-bulk-100.json";
+import worldRpcLogs1000 from "../../../test-data/world-logs-bulk-1000.json";
 
 describe.each([
   { numRecords: 10, logs: worldRpcLogs10 },
